modules/API_Voice/requirements.txt
pyaudio
torch
requests
edge-tts
pydub
websocket-client
numpy